chore: merge test_xgboost_mb.py and test_model_builders.py
